Saving guests data and reservations data to database

+++ src/main/java/pl/emilweglowski/domain/guest/GuestService.java
@@ -39,11 +39,11 @@ public class GuestService {
         this.repository.readAll();
     }
 
-    public void removeGuest(int id) {
+    public void removeGuest(long id) {
         this.repository.remove(id);
     }
 
-    public void editGuest(int id, String firstName, String lastName, int age, boolean isMale) {
+    public void editGuest(long id, String firstName, String lastName, int age, boolean isMale) {
         Gender gender = Gender.FEMALE;
         if (isMale){
             gender = Gender.MALE;
@@ -51,7 +51,7 @@ public class GuestService {
         this.repository.edit(id, firstName, lastName, age, gender);
     }
 
-    public Guest getGuestById(int guestId) {
+    public Guest getGuestById(long guestId) {
         return this.repository.getById(guestId);
     }
