+++ src/main/java/pl/emilweglowski/domain/reservation/ReservationService.java
@@ -27,7 +27,7 @@ public class ReservationService {
         return instance;
     }
 
-    public Reservation createNewReservation(LocalDate from, LocalDate to, int roomId, int guestId) throws IllegalArgumentException {
+    public Reservation createNewReservation(LocalDate from, LocalDate to, long roomId, long guestId) throws IllegalArgumentException {
 
         //TODO: handle null room
         Room room = this.roomService.getRoomById(roomId);
@@ -55,7 +55,6 @@ public class ReservationService {
     public List<ReservationDTO> getReservationsAsDTO() {
         List<ReservationDTO> result = new ArrayList<>();
         List<Reservation> allReservations = repository.getAll();
-
         for(Reservation reservation : allReservations) {
             ReservationDTO dto = reservation.generateDTO();
             result.add(dto);
@@ -63,7 +62,7 @@ public class ReservationService {
         return result;
     }
 
-    public void removeReservation(int id) {
+    public void removeReservation(long id) {
         this.repository.remove(id);
     }
 }
